Move tokenAtPos to position module
