+++ frontend/src/metabase/querying/expressions/suggestions/functions.ts
@@ -4,6 +4,7 @@ import type * as Lib from "metabase-lib";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
 
 import { getSupportedClauses } from "../clause";
+import { tokenAtPos } from "../position";
 import {
   GROUP,
   LOGICAL_AND,
@@ -19,7 +20,6 @@ import {
   fuzzyMatcher,
   isFieldReference,
   isIdentifier,
-  tokenAtPos,
 } from "./util";
 
 export type Options = {
