+++ frontend/src/metabase/querying/expressions/suggestions/util.ts
@@ -2,15 +2,7 @@ import { snippetCompletion } from "@codemirror/autocomplete";
 import Fuse from "fuse.js";
 import _ from "underscore";
 
-import {
-  CALL,
-  END_OF_INPUT,
-  FIELD,
-  IDENTIFIER,
-  STRING,
-  type Token,
-  lexify,
-} from "../pratt";
+import { CALL, FIELD, IDENTIFIER, type Token } from "../pratt";
 import type { MBQLClauseFunctionConfig } from "../types";
 
 import type { ExpressionSuggestion } from "./types";
@@ -77,31 +69,6 @@ export const fuzzyMatcher = _.memoize(function ({
       });
   };
 }, JSON.stringify);
-
-export function tokenAtPos(source: string, pos: number): Token | null {
-  const tokens = lexify(source);
-
-  const idx = tokens.findIndex(
-    (token) => token.start <= pos && token.end >= pos,
-  );
-  if (idx === -1) {
-    return null;
-  }
-
-  const token = tokens[idx];
-  const prevToken = tokens[idx - 1];
-
-  if (token.type === END_OF_INPUT) {
-    return null;
-  }
-
-  if (prevToken && prevToken.type === STRING && prevToken.length === 1) {
-    // dangling single- or double-quote
-    return null;
-  }
-
-  return token;
-}
 
 export function content(source: string, token: Token): string {
   return source.slice(token.pos, token.pos + token.length);
